refactor: RacingCarInput 클래스명 목적에 맞게 명칭 변경
